requirements.txt
flask==3.0.3
pymysql==1.1.0
cryptography==42.0.7
flask
pymysql
boto3
requests
python-dotenv
